+++ src/app/loadout-builder/process-worker/process.ts
@@ -13,11 +13,7 @@ import {
   StatFilters,
   StatRanges,
 } from '../types';
-import {
-  canTakeSlotIndependentMods,
-  generateProcessModPermutations,
-  sortProcessModsOrItems,
-} from './process-utils';
+import { pickAndAssignSlotIndependentMods, precalculateStructures } from './process-utils';
 import { SetTracker } from './set-tracker';
 import {
   LockedProcessMods,
@@ -46,6 +42,8 @@ export function process(
   statFilters: StatFilters,
   /** Ensure every set includes one exotic */
   anyExotic: boolean,
+  /** Use stat mods to hit stat minimums */
+  autoStatMods: boolean,
   onProgress: (remainingTime: number) => void
 ): {
   sets: ProcessArmorSet[];
@@ -119,14 +117,12 @@ export function process(
     }
   }
 
-  const generalModsPermutations = generateProcessModPermutations(
-    generalMods.sort(sortProcessModsOrItems)
-  );
-  const combatModPermutations = generateProcessModPermutations(
-    combatMods.sort(sortProcessModsOrItems)
-  );
-  const activityModPermutations = generateProcessModPermutations(
-    activityMods.sort(sortProcessModsOrItems)
+  const precalculatedInfo = precalculateStructures(
+    generalMods,
+    combatMods,
+    activityMods,
+    autoStatMods,
+    statOrder
   );
   const hasMods = Boolean(combatMods.length || activityMods.length || generalMods.length);
 
@@ -228,7 +224,7 @@ export function process(
               const tier = tiers[index];
               const filter = statFiltersInStatOrder[index];
               if (!filter.ignored) {
-                if (tier > filter.max || tier < filter.min) {
+                if (tier > filter.max) {
                   statRangeExceeded = true;
                 }
                 totalTier += tier;
@@ -248,20 +244,49 @@ export function process(
 
             const armor = [helm, gaunt, chest, leg, classItem];
 
+            const neededStats = [0, 0, 0, 0, 0, 0];
+            let needSomeStats = false;
+
+            // Check in which stats we're lacking
+            for (let index = 0; index < 6; index++) {
+              const value = Math.min(Math.max(stats[index], 0), 100);
+              const filter = statFiltersInStatOrder[index];
+
+              if (!filter.ignored) {
+                const neededValue = filter.min * 10 - value;
+                if (neededValue > 0) {
+                  neededStats[index] = neededValue;
+                  needSomeStats = true;
+                }
+              }
+            }
+
+            if (needSomeStats && !autoStatMods) {
+              numStatRangeExceeded++;
+              continue;
+            }
+
+            let statMods: number[] = [];
             // For armour 2 mods we ignore slot specific mods as we prefilter items based on energy requirements
             // TODO: this isn't a big part of the overall cost of the loop, but we could consider trying to slot
             // mods at every level (e.g. just helmet, just helmet+arms) and skipping this if they already fit.
-            if (
-              hasMods &&
-              !canTakeSlotIndependentMods(
-                generalModsPermutations,
-                combatModPermutations,
-                activityModPermutations,
-                armor
-              )
-            ) {
-              numCantSlotMods++;
-              continue;
+            if (hasMods || needSomeStats) {
+              const modPickResult = pickAndAssignSlotIndependentMods(
+                precalculatedInfo,
+                armor,
+                needSomeStats ? neededStats : undefined
+              );
+
+              switch (modPickResult) {
+                case 'cannot_hit_stats':
+                  numStatRangeExceeded++;
+                  continue;
+                case 'mods_dont_fit':
+                  numCantSlotMods++;
+                  continue;
+                default:
+                  statMods = modPickResult.modHashes;
+              }
             }
 
             // Calculate the "tiers string" here, since most sets don't make it this far
@@ -270,7 +295,6 @@ export function process(
             // than comparing stat arrays element by element
             let tiersString = '';
             for (let index = 0; index < 6; index++) {
-              const value = Math.min(Math.max(stats[index], 0), 100);
               const tier = tiers[index];
               // Make each stat exactly one code unit so the string compares correctly
               const filter = statFiltersInStatOrder[index];
@@ -281,6 +305,7 @@ export function process(
 
               // Track the stat ranges of sets that made it through all our filters
               const range = statRangesFilteredInStatOrder[index];
+              const value = stats[index];
               if (value > range.max) {
                 range.max = value;
               }
@@ -290,7 +315,7 @@ export function process(
             }
 
             numValidSets++;
-            setTracker.insert(totalTier, tiersString, armor, stats);
+            setTracker.insert(totalTier, tiersString, armor, stats, statMods);
           }
         }
       }
@@ -333,13 +358,19 @@ export function process(
       numNoExotic,
     }
   );
+  infoLog('loadout optimizer', 'auto stat mods', {
+    cacheHits: precalculatedInfo.cache.cacheHits,
+    cacheMisses: precalculatedInfo.cache.cacheMisses,
+    cacheSuccesses: precalculatedInfo.cache.cacheSuccesses,
+  });
 
-  const sets = finalSets.map(({ armor, stats }) => ({
+  const sets = finalSets.map(({ armor, stats, statMods }) => ({
     armor: armor.map((item) => item.id),
     stats: statOrder.reduce((statObj, statHash, i) => {
       statObj[statHash] = stats[i];
       return statObj;
     }, {}) as ArmorStats,
+    statMods,
   }));
 
   return {