+++ src/app/search/d2-known-values.ts
@@ -242,6 +242,7 @@ export const VENDORS = {
   VAULT: 1037843411,
   XUR: 2190858386,
   WAR_TABLE_UPGRADES_RISEN: 3950870173,
+  STAR_CHART_UPGRADES_PLUNDER: 3004285529,
   /**
    * this has always been named "The Gate Lord's Eye" from season 8,
    * but every season this vendor is updated with the new contents
@@ -297,13 +298,14 @@ export type ItemTierName = keyof typeof D2ItemTiers & string;
 export const energyCapacityTypeNames = Object.values(energyNamesByEnum);
 
 export const breakerTypes = {
-  barrier: BreakerTypeHashes.ShieldPiercing,
-  antibarrier: BreakerTypeHashes.ShieldPiercing,
-  shieldpiercing: BreakerTypeHashes.ShieldPiercing,
-  overload: BreakerTypeHashes.Disruption,
-  disruption: BreakerTypeHashes.Disruption,
-  unstoppable: BreakerTypeHashes.Stagger,
-  stagger: BreakerTypeHashes.Stagger,
+  any: [BreakerTypeHashes.Stagger, BreakerTypeHashes.Disruption, BreakerTypeHashes.ShieldPiercing],
+  barrier: [BreakerTypeHashes.ShieldPiercing],
+  antibarrier: [BreakerTypeHashes.ShieldPiercing],
+  shieldpiercing: [BreakerTypeHashes.ShieldPiercing],
+  overload: [BreakerTypeHashes.Disruption],
+  disruption: [BreakerTypeHashes.Disruption],
+  unstoppable: [BreakerTypeHashes.Stagger],
+  stagger: [BreakerTypeHashes.Stagger],
 };
 
 export const modsWithConditionalStats = {