+++ src/app/search/search-filters/sockets.tsx
@@ -6,7 +6,11 @@ import {
   modSlotTags,
   modTypeTags,
 } from 'app/utils/item-utils';
-import { countEnhancedPerks, getSocketsByCategoryHash } from 'app/utils/socket-utils';
+import {
+  countEnhancedPerks,
+  getIntrinsicArmorPerkSocket,
+  getSocketsByCategoryHash,
+} from 'app/utils/socket-utils';
 import { DestinyItemSubType, DestinyRecordState } from 'bungie-api-ts/destiny2';
 import craftingMementos from 'data/d2/crafting-mementos.json';
 import {
@@ -183,6 +187,12 @@ const socketFilters: FilterDefinition[] = [
             socket.plugged.plugDef.perks.length
         )
       ),
+  },
+  {
+    keywords: 'armorintrinsic',
+    description: tl('Filter.ArmorIntrinsic'),
+    destinyVersion: 2,
+    filter: () => (item: DimItem) => Boolean(!item.isExotic && getIntrinsicArmorPerkSocket(item)),
   },
   {
     keywords: 'holdsmod',