+++ src/app/vendors/VendorItems.tsx
@@ -54,6 +54,9 @@ function itemSort(vendorHash: number, category: string) {
     // Purchasing an upgrade from the vendor swaps it out with a different item
     // 10 positions later in the array.
     return compareBy<VendorItem>((item) => item.key - (item.owned ? 10 : 0));
+  } else if (vendorHash === VENDORS.STAR_CHART_UPGRADES_PLUNDER) {
+    // Basically the same thing
+    return compareBy<VendorItem>((item) => item.key - (item.owned ? 21 : 0));
   } else {
     return chainComparator<VendorItem>(compareBy(vendorItemIndex));
   }