+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `is:armorintrinsic` to find Artifice Armor, armor with seasonal perks, etc.
 * Compare suggestion buttons now offer comparison to similar armor intrinsics.
+* Added perks to Light.gg links. See your weapon's popularity rating without having to reselect it's perks.
 
 ## 7.38.0 <span class="changelog-date">(2022-10-09)</span>
 
@@ -12,7 +13,6 @@
 ## 7.37.0 <span class="changelog-date">(2022-10-02)</span>
 
 * Add `foundry` search term. Try `foundry:hakke` for all your items brought to you by Hakke.
-* Added perks to Light.gg links. See your weapon's popularity rating without having to reselect it's perks.
 
 ## 7.36.0 <span class="changelog-date">(2022-09-25)</span>